--- Large Refactor ---
Working
