+++ src/main.rs
@@ -4,128 +4,21 @@
 
 // External crates
 
-use async_osc::{prelude::*, OscPacket, OscSocket, OscType, Result};
+use async_osc::{prelude::*, OscPacket, OscType, Result};
 use async_std::{
     stream::StreamExt,
     task::{self},
     sync::Arc,
 
 };
-use configparser::ini::Ini;
 use lazy_static::lazy_static;
 use std::{sync::Mutex, time::{Duration, Instant}};
 use std::sync::atomic::{AtomicBool, Ordering};
 
-
-// Banner
-fn banner_txt(){
-    // https://fsymbols.com/generators/carty/
-    println!("");
-    println!("  ██████  ██  ██████   ██████  ██      ███████     ████████ ███████  ██████ ██   ██ ");
-    println!(" ██       ██ ██       ██       ██      ██             ██    ██      ██      ██   ██ ");
-    println!(" ██   ███ ██ ██   ███ ██   ███ ██      █████          ██    █████   ██      ███████ ");
-    println!(" ██    ██ ██ ██    ██ ██    ██ ██      ██             ██    ██      ██      ██   ██ ");
-    println!("  ██████  ██  ██████   ██████  ███████ ███████        ██    ███████  ██████ ██   ██ ");
-    println!("");
-    println!(" █▀█ █▀ █▀▀   █▀█ █▀█ █ █ ▀█▀ █▀▀ █▀█");
-    println!(" █▄█ ▄█ █▄▄   █▀▄ █▄█ █▄█  █  ██▄ █▀▄");
-                                                                                
-}
-
-fn load_config() -> (
-    String, // headpat_device_ip
-    String, // headpat_device_port
-    f32,    // min_speed_float
-    f32,    // max_speed_float
-    f32,    // speed_scale_float
-    String, // port_rx
-    String, // proximity_parameter_address
-    String, // max_speed_parameter_address
-    f32,    // Max Speed Low Limit
-    ) {
-    let mut config = Ini::new();
-
-    match config.load("./config.ini") {
-        Err(why) => panic!("{}", why),
-        Ok(_) => {}
-    }
-    const MAX_SPEED_LOW_LIMIT_CONST: f32 = 0.05;
-
-    let headpat_device_ip   = config.get("Setup", "device_ip").unwrap();
-    let headpat_device_port = "8888".to_string();
-    let min_speed           = config.get("Haptic_Config", "min_speed").unwrap();
-    let min_speed_float     = min_speed.parse::<f32>().unwrap() / 100.0;
-    let max_speed           = config.get("Haptic_Config", "max_speed").unwrap();
-    let max_speed_float     = max_speed.parse::<f32>().unwrap() / 100.0; 
-    let max_speed_low_limit = MAX_SPEED_LOW_LIMIT_CONST;
-    let max_speed_float     = max_speed_float.max(max_speed_low_limit);
-    let speed_scale         = config.get("Haptic_Config", "max_speed_scale").unwrap();
-    let speed_scale_float   = speed_scale.parse::<f32>().unwrap() / 100.0;
-    let port_rx             = config.get("Setup", "port_rx").unwrap();
-
-    let proximity_parameter_address = config
-        .get("Setup", "proximity_parameter")
-        .unwrap_or_else(|| "/avatar/parameters/proximity_01".into());
-    let max_speed_parameter_address = config
-        .get("Setup", "max_speed_parameter")
-        .unwrap_or_else(|| "/avatar/parameters/max_speed".into());
-
-    println!("\n");
-    banner_txt();
-    println!("\n");
-    println!(" Haptic Device: {}:{}", headpat_device_ip, headpat_device_port);
-    println!(" Listening for OSC on port: {}", port_rx);
-    println!("\n Vibration Configuration");
-    println!(" Min Speed: {}%", min_speed);
-    println!(" Max Speed: {:?}%", max_speed_float * 100.0);
-    println!(" Scale Factor: {}%", speed_scale);
-    println!("\nWaiting for pats...");
-
-    (
-        headpat_device_ip,
-        headpat_device_port,
-        min_speed_float,
-        max_speed_float,
-        speed_scale_float,
-        port_rx,
-        proximity_parameter_address,
-        max_speed_parameter_address,
-        max_speed_low_limit,
-    )
-}
-
-// Make it easy to see prox when looking at router
-fn proximity_graph(proximity_signal: f32) -> String {
-    let num_dashes = (proximity_signal * 10.0) as usize;
-    let graph = "-".repeat(num_dashes) + ">";
-
-    graph
-}
-
-fn print_speed_limit(headpat_max_rx: f32) {
-    let headpat_max_rx_print = (headpat_max_rx * 100.0).round() as i32;
-    let max_meter = match headpat_max_rx_print {
-        91..=i32::MAX => "!!! SO MUCH !!!",
-        76..=90 => "!! ",
-        51..=75 => "!  ",
-        _ => "   ",
-    };
-    println!("Speed Limit: {}% {}", headpat_max_rx_print, max_meter);
-}
-
-// Pat Processor
-const MOTOR_SPEED_SCALE: f32 = 0.66; // Overvolt   Here, OEM config 0.66 going higher than this value will reduce your vibrator motor life
-fn process_pat(proximity_signal: f32, max_speed: f32, min_speed: f32, speed_scale: f32) -> i32 {
-    let graph_str = proximity_graph(proximity_signal);
-    let headpat_tx = (((max_speed - min_speed) * proximity_signal + min_speed) * MOTOR_SPEED_SCALE * speed_scale * 255.0).round() as i32;
-    let proximity_signal = format!("{:.2}", proximity_signal);
-    let max_speed = format!("{:.2}", max_speed);
-    eprintln!("Prox: {:5} Motor Tx: {:3}  Max Speed: {:5} |{:11}|", proximity_signal, headpat_tx, max_speed, graph_str);
-    
-    headpat_tx
-}
-
-
+// Modules
+mod data_processing;
+mod config;
+mod socket_setup;
 
 
 
@@ -135,10 +28,8 @@ lazy_static! {
 }
 
 async fn osc_timeout(device_ip: &str) -> Result<()> {
-    let tx_socket_address = create_socket_address(device_ip, "8888");
-    let mut tx_socket = setup_tx_socket(tx_socket_address).await?;
-
     // If no new osc signal is Rx for 5s, will send stop packets
+    // This loop can be used to implement Kays 'Soft Pat'
     loop {
         task::sleep(Duration::from_secs(1)).await;
         let elapsed_time = Instant::now().duration_since(*LAST_SIGNAL_TIME.lock().unwrap());
@@ -153,8 +44,6 @@ async fn osc_timeout(device_ip: &str) -> Result<()> {
         }
     }
 }
-
-
 
 
 /*
@@ -176,10 +65,6 @@ async fn start(running: Arc<AtomicBool>, running_mutex: Arc<Mutex<()>>) -> Resul
     
     Ok(())
 }
-
-
-
-
 
 
 async fn worker(running: Arc<AtomicBool>, device_ip: &str) -> Result<()> {
@@ -211,24 +96,7 @@ async fn stop(
 }
 
 
-// Tx & Rx Socket Setup
 
-fn create_socket_address(host: &str, port: &str) -> String {
-    let address_parts = vec![host, port];
-    address_parts.join(":")
-}
-
-async fn setup_rx_socket(port: std::string::String) -> Result<OscSocket> {
-    let rx_socket_address = create_socket_address("127.0.0.1", &port.to_string());
-    let rx_socket = OscSocket::bind(rx_socket_address).await?;
-    Ok(rx_socket)
-}
-
-async fn setup_tx_socket(address: std::string::String) -> Result<OscSocket> {
-    let tx_socket = OscSocket::bind("0.0.0.0:0").await?;
-    tx_socket.connect(address).await?;
-    Ok(tx_socket)
-}
 
 
 // OSC Address Setup
@@ -238,8 +106,8 @@ const TX_OSC_MOTOR_ADDRESS: &str = "/avatar/parameters/motor";
 
 async fn send_data(device_ip: &str, address: &str, value: i32) -> Result<()> {
     println!("Sending Value:{} to IP: {}", value, device_ip);
-    let tx_socket_address = create_socket_address(device_ip, "8888"); // -------------------------------------------------- FIX
-    let mut tx_socket = setup_tx_socket(tx_socket_address.clone()).await?;
+    let tx_socket_address = socket_setup::create_socket_address(device_ip, "8888"); // ------------------- Port to Send OSC Data Too
+    let tx_socket = socket_setup::setup_tx_socket(tx_socket_address.clone()).await?;
     tx_socket.connect(tx_socket_address).await?;
     tx_socket.send((address, (value,))).await?;
     Ok(())
@@ -252,7 +120,7 @@ async fn main() -> Result<()> {
      
     // Import Config 
     let (headpat_device_ip,
-        headpat_device_port,
+        _headpat_device_port,
         min_speed,
         mut max_speed,
         speed_scale,
@@ -261,10 +129,10 @@ async fn main() -> Result<()> {
         max_speed_parameter_address,
         max_speed_low_limit,
 
-    ) = load_config();
+    ) = config::load_config();
 
     // Rx/Tx Socket Setup
-    let mut rx_socket = setup_rx_socket(port_rx).await?;
+    let mut rx_socket = socket_setup::setup_rx_socket(port_rx).await?;
 
  
     // Timeout
@@ -296,7 +164,7 @@ async fn main() -> Result<()> {
 
                 // Max Speed Setting
                 if address == max_speed_parameter_address {
-                    print_speed_limit(value);
+                    data_processing::print_speed_limit(value);
                     max_speed = value.max(max_speed_low_limit);
                 }
                 
@@ -321,9 +189,11 @@ async fn main() -> Result<()> {
                         }
 
                     } else {
-                        // Process Pat signal to send to Device 
-                        let motor_speed_tx = process_pat(value, max_speed, min_speed, speed_scale);
-                        send_data(&headpat_device_ip, TX_OSC_MOTOR_ADDRESS, motor_speed_tx).await?;
+
+                        //let motor_speed_tx = data_processing::process_pat(value, max_speed, min_speed, speed_scale);
+                        send_data(&headpat_device_ip,
+                            TX_OSC_MOTOR_ADDRESS,
+                            data_processing::process_pat(value, max_speed, min_speed, speed_scale)).await?;
                         
 
                     }
