+++ src/socket_setup.rs
@@ -1,0 +1,21 @@
+// Tx & Rx Socket Setup
+// External crates
+
+use async_osc::{ OscSocket, Result};
+
+pub(crate) fn create_socket_address(host: &str, port: &str) -> String {
+    let address_parts = vec![host, port];
+    address_parts.join(":")
+}
+
+pub(crate) async fn setup_rx_socket(port: std::string::String) -> Result<OscSocket> {
+    let rx_socket_address = create_socket_address("127.0.0.1", &port.to_string());
+    let rx_socket = OscSocket::bind(rx_socket_address).await?;
+    Ok(rx_socket)
+}
+
+pub(crate) async fn setup_tx_socket(address: std::string::String) -> Result<OscSocket> {
+    let tx_socket = OscSocket::bind("0.0.0.0:0").await?;
+    tx_socket.connect(address).await?;
+    Ok(tx_socket)
+}
